增加图片选择界面和util

+++ app/src/main/java/com/shiqkuangsan/mycustomviews/utils/ChosePicUtil.java
@@ -1,13 +1,11 @@
 package com.shiqkuangsan.mycustomviews.utils;
 
 /**
- * Created by dell on 2016/9/6.
+ * Created by shiqkuangsan on 2016/9/6.
  */
 
 import android.app.Activity;
-import android.app.LoaderManager;
 import android.content.ContentResolver;
-import android.content.Context;
 import android.content.Intent;
 import android.database.Cursor;
 import android.graphics.Bitmap;
@@ -15,17 +13,18 @@ import android.graphics.BitmapFactory;
 import android.net.Uri;
 import android.os.Bundle;
 import android.os.Environment;
+import android.os.Parcelable;
 import android.provider.MediaStore;
 import android.widget.Toast;
-
-import com.shiqkuangsan.mycustomviews.ui.PicChoserActivity;
 
 import java.io.File;
 import java.io.FileNotFoundException;
 
 /**
  * 选择图片的工具类
- * 1.在你的两个选择按钮调用startActivityFor方法传参选择获取方式
+ * 1. 在你的两个选择按钮分别调用startActivityFor()方法传入不同参书选择获取方式进入对应界面
+ * 2. 在你的onActivityResult()方法处调用getBitmapFromResult()方法,将会获取到选择的图片的bitmap对象,直接使用即可.
+ * 如果调用getPathFromResult()方法,将会获取到图片的路径(后者不支持裁剪且从相册获取使用过之后若想删除缓存需要调用deleteTemp())
  */
 public class ChosePicUtil {
 
@@ -50,7 +49,7 @@ public class ChosePicUtil {
     public static final int REQUEST_CODE_CAMERA = 200;
 
     /**
-     * 从相机获取的请求码
+     * 进入裁剪界面的请求码
      */
     public static final int REQUEST_CODE_CROP = 300;
 
@@ -63,7 +62,7 @@ public class ChosePicUtil {
     /**
      * 根据不同匹配码确定是从图库选择还是相机获取,并启动相应activity
      *
-     * @param MATCHING_CODE 匹配码   MATCHING_CODE_GALLERY - 图库获取, MATCHING_CODE_CAMERA - 相机获取
+     * @param MATCHING_CODE 匹配码, MATCHING_CODE_GALLERY - 图库获取, MATCHING_CODE_CAMERA - 相机获取
      */
     public static void startActivityFor(int MATCHING_CODE, Activity activity) {
         switch (MATCHING_CODE) {
@@ -91,16 +90,18 @@ public class ChosePicUtil {
     }
 
     /**
-     * 根据返回的结果进行不同的解析
+     * 根据返回的结果进行不同的解析,获取bitmap对象
      *
-     * @param requestCode 请求码
-     * @param resultCode  结果码
-     * @param data        返回数据(里边封装这一个uri)
-     * @param activity    当前activity
-     * @param needCrop    获取完之后是否需要裁剪? true - 进入裁剪界面
+     * @param requestCode    请求码
+     * @param resultCode     结果码
+     * @param data           返回数据(里边封装这一个uri)
+     * @param activity       当前activity
+     * @param needCrop       获取完之后是否需要裁剪? true - 进入裁剪界面
+     * @param needDeleteTemp 如果是拍照获取,获取完之后是否需要删除缓存照片? true - 删除
      * @return Bitmap对象, null - 解析失败
      */
-    public static Bitmap getBitmapFromResult(int requestCode, int resultCode, Intent data, Activity activity, boolean needCrop) {
+    public static Bitmap getBitmapFromResult(int requestCode, int resultCode, Intent data,
+                                             Activity activity, boolean needCrop, boolean needDeleteTemp) {
         if (resultCode != Activity.RESULT_OK)
             return null;
 
@@ -134,6 +135,12 @@ public class ChosePicUtil {
                     else
                         try {
                             bitmap = BitmapFactory.decodeStream(resolver.openInputStream(Uri.fromFile(tempFile)));
+                            if (needDeleteTemp)
+                                try {
+                                    tempFile.delete();
+                                } catch (Exception e) {
+                                    e.printStackTrace();
+                                }
                         } catch (FileNotFoundException e) {
                             e.printStackTrace();
                         }
@@ -146,14 +153,16 @@ public class ChosePicUtil {
                 if (data != null) {
                     bitmap = data.getParcelableExtra("data");
                 }
-                try {
-                    // 将临时文件删除
-                    tempFile.delete();
-                } catch (Exception e) {
-                    e.printStackTrace();
-                }
+                if (needDeleteTemp)
+                    try {
+                        tempFile.delete();
+                    } catch (Exception e) {
+                        e.printStackTrace();
+                    }
                 break;
+
         }
+
         return bitmap;
 
     }
@@ -175,7 +184,6 @@ public class ChosePicUtil {
         // 裁剪后输出图片的尺寸大小
         intent.putExtra("outputX", 250);
         intent.putExtra("outputY", 250);
-
         intent.putExtra("outputFormat", "JPEG");// 图片格式
         intent.putExtra("noFaceDetection", true);// 取消人脸识别
         intent.putExtra("return-data", true);
@@ -184,14 +192,16 @@ public class ChosePicUtil {
 
 
     /**
-     * @param requestCode
-     * @param resultCode
-     * @param data
-     * @param activity
-     * @param needCrop
-     * @return
+     * 根据返回的结果进行不同的解析,获取文件路径
+     * (如果调用此方法,拍照获取路径后需要手动调用deleteTemp方法删除缓存照片)
+     *
+     * @param requestCode 请求码
+     * @param resultCode  结果码
+     * @param data        返回数据(里边封装这一个uri)
+     * @param activity    activity
+     * @return 文件路径的字符串, "" / null - 获取失败
      */
-    public static String getStringFromResult(int requestCode, int resultCode, Intent data, Activity activity, boolean needCrop) {
+    public static String getPathFromResult(int requestCode, int resultCode, Intent data, Activity activity) {
 
         if (resultCode != Activity.RESULT_OK)
             return null;
@@ -201,47 +211,53 @@ public class ChosePicUtil {
             // 从图库返回
             case REQUEST_CODE_GALLERY:
                 if (data != null) {
-                    // 根据返回的数据利用内容解决者解析出bitmap对象
+                    // 根据返回的数据利用内容解决者解析出文件路径
                     Uri uri = data.getData();
-                    if (needCrop)
-                        crop(uri, activity);
-                    else
-                        imgPath = setImagePath(uri, activity);
+                    imgPath = getImagePath(uri, activity);
                 }
                 break;
 
             // 从相机返回
             case REQUEST_CODE_CAMERA:
-                if (Environment.getExternalStorageState().equals(Environment.MEDIA_MOUNTED)) {
-                    if (needCrop)
-                        crop(Uri.fromFile(tempFile), activity);
-                    else
-                        imgPath = setImagePath(Uri.fromFile(tempFile), activity);
-                }
-                break;
-
-            // 从剪切界面返回
-            case REQUEST_CODE_CROP:
-                if (data != null) {
-                    Bundle extras = data.getExtras();
-                }
-                try {
-                    // 将临时文件删除
-                    tempFile.delete();
-                } catch (Exception e) {
-                    e.printStackTrace();
-                }
+                if (Environment.getExternalStorageState().equals(Environment.MEDIA_MOUNTED))
+                    imgPath = tempFile.getPath();
                 break;
         }
+        // 这里如果删除了临时文件,返回的imgPath也没有意义了,因为文件已经没了,所以提供方法
         return imgPath;
 
     }
 
-    public static String setImagePath(Uri uri, Activity activity) {
+    /**
+     * 根据图片的uri返回图片的路径
+     *
+     * @param uri      资源标识uri
+     * @param activity activity界面
+     * @return 图片的路径
+     */
+    public static String getImagePath(Uri uri, Activity activity) {
         String[] proj = {MediaStore.Images.Media.DATA};
-        Cursor cursor = activity.managedQuery(uri, proj, null, null, null);
-        int column_index = cursor.getColumnIndexOrThrow(MediaStore.Images.Media.DATA);
-        cursor.moveToFirst();
-        return cursor.getString(column_index);
+        // 利用内容提供者去查询图片
+        ContentResolver resolver = activity.getContentResolver();
+        Cursor cursor = resolver.query(uri, proj, null, null, null);
+        String path = "";
+        if (cursor != null) {
+            cursor.moveToFirst();
+            path = cursor.getString(cursor.getColumnIndex(MediaStore.Images.Media.DATA));
+            cursor.close();
+        }
+        return path;
+    }
+
+    /**
+     * 如果是拍照获取的图片,调用此方法删除缓存照片
+     *
+     * @return 删除是否成功
+     */
+    public static boolean deleteTemp() {
+        boolean result = false;
+        if (tempFile != null)
+            result = tempFile.delete();
+        return result;
     }
 }
